Extracted Main Menu into its own package
